+++ main.py
@@ -32,6 +32,7 @@ class Main(KytosNApp):  # pylint: disable=too-many-public-methods
         """Initialize the NApp's links list."""
         self.links = {}
         self.store_items = {}
+        self.intf_available_tags = {}
         self.link_up_timer = getattr(settings, 'LINK_UP_TIMER',
                                      DEFAULT_LINK_UP_TIMER)
 
@@ -109,6 +110,21 @@ class Main(KytosNApp):  # pylint: disable=too-many-public-methods
         return {'links': {link.id: link.as_dict() for link in
                           self.links.values()}}
 
+    def _get_links_dict_with_tags(self):
+        """Return a dict with the known links with stored available_tags."""
+        links = {}
+        for link in self.links.values():
+            link_dict = link.as_dict()
+            endpoint_a = link_dict["endpoint_a"]
+            endpoint_b = link_dict["endpoint_b"]
+            for endpoint in (endpoint_a, endpoint_b):
+                if endpoint["id"] in self.intf_available_tags:
+                    endpoint["available_tags"] = self.intf_available_tags[
+                        endpoint["id"]
+                    ]
+            links[link.id] = link_dict
+        return {"links": links}
+
     def _get_topology_dict(self):
         """Return a dictionary with the known topology."""
         return {'topology': {**self._get_switches_dict(),
@@ -124,6 +140,14 @@ class Main(KytosNApp):  # pylint: disable=too-many-public-methods
             if interface in (link.endpoint_a, link.endpoint_b):
                 return link
         return None
+
+    @staticmethod
+    def _load_intf_available_tags(interface, intf_dict):
+        """Load interface available tags given its dict."""
+        has_available_tags = "available_tags" in intf_dict
+        if has_available_tags:
+            interface.set_available_tags(intf_dict["available_tags"])
+        return has_available_tags
 
     def _load_link(self, link_att):
         dpid_a = link_att['endpoint_a']['switch']
@@ -154,6 +178,15 @@ class Main(KytosNApp):  # pylint: disable=too-many-public-methods
         interface_b.update_link(link)
         interface_a.nni = True
         interface_b.nni = True
+        for interface, interface_dict in (
+            (interface_a, link_att["endpoint_a"]),
+            (interface_b, link_att["endpoint_b"]),
+        ):
+            if self._load_intf_available_tags(interface, interface_dict):
+                log.info(
+                    f"Loaded {len(interface.available_tags)}"
+                    f" available tags for {interface.id}"
+                )
         self.update_instance_metadata(link)
 
     def _load_switch(self, switch_id, switch_att):
@@ -523,6 +556,25 @@ class Main(KytosNApp):  # pylint: disable=too-many-public-methods
         self.notify_metadata_changes(link, 'removed')
         return jsonify("Operation successful"), 200
 
+    @listen_to("kytos/.*.link_available_tags")
+    def on_link_available_tags(self, event):
+        """Handle on_link_available_tags."""
+        with self._links_lock:
+            self.handle_on_link_available_tags(event.content.get("link"))
+
+    def handle_on_link_available_tags(self, link):
+        """Handle on_link_available_tags."""
+        if link.id not in self.links:
+            return
+        endpoint_a = self.links[link.id].endpoint_a
+        endpoint_b = self.links[link.id].endpoint_b
+        tags_a = [tag.value for tag in endpoint_a.available_tags]
+        tags_b = [tag.value for tag in endpoint_b.available_tags]
+        self.intf_available_tags[endpoint_a.id] = tags_a
+        self.intf_available_tags[endpoint_b.id] = tags_b
+
+        self.save_status_on_storehouse()
+
     @listen_to('.*.switch.(new|reconnected)')
     def on_new_switch(self, event):
         """Create a new Device on the Topology.
@@ -566,9 +618,20 @@ class Main(KytosNApp):  # pylint: disable=too-many-public-methods
         created event again and it can be belong to a link.
         """
         interface = event.content['interface']
-        self.notify_topology_update()
+
+        with self._links_lock:
+            if interface.id in self.intf_available_tags:
+                available_tags = self.intf_available_tags[interface.id]
+                interface_dict = {"available_tags": available_tags}
+                if self._load_intf_available_tags(interface, interface_dict):
+                    log.info(
+                        f"Loaded {len(interface.available_tags)}"
+                        f" available tags for {interface.id}"
+                    )
+
         self.update_instance_metadata(interface)
         self.handle_interface_link_up(interface)
+        self.notify_topology_update()
 
     @listen_to('.*.switch.interface.created')
     def on_interface_created(self, event):
@@ -765,7 +828,7 @@ class Main(KytosNApp):  # pylint: disable=too-many-public-methods
         with self._lock:
             status = self._get_switches_dict()
             status['id'] = 'network_status'
-            status.update(self._get_links_dict())
+            status.update(self._get_links_dict_with_tags())
             self.storehouse.save_status(status)
 
     def notify_switch_enabled(self, dpid):